UserTest Done with link from gallery Need to finish to bind with the other entities

+++ routes/api.php
@@ -16,9 +16,12 @@ use Illuminate\Support\Facades\Route;
 Route::get('/ping', 'Api\PingController@ping');
 
 Route::group(["namespace" => 'Api', 'prefix' => 'user'], function () {
-    Route::post('', 'UserController@store');
+    Route::post('', 'UserController@store') ;
+    Route::post('/member', 'UserController@storeMember')->middleware('permission:store member');
     Route::get('', 'UserController@index');
     Route::get('/self', 'UserController@showByToken');
+    Route::get('/self/permissions', 'UserController@getOwnPermissions');
+    Route::post('/role/{user}', 'UserController@updateRole')->middleware('permission:update role');
     Route::get('/{user}', 'UserController@show');
     Route::patch('', 'UserController@update');
     Route::delete('', 'UserController@destroy');
@@ -33,19 +36,19 @@ Route::group(["namespace" => 'Api', 'prefix' => 'gallery', 'middleware' => 'auth
 });
 
 Route::group(["namespace" => 'Api', 'prefix' => 'artist', 'middleware' => 'auth:api'], function () {
-    Route::post('', 'ArtistController@store');
-    Route::get('', 'ArtistController@index');
-    Route::get('/{artist}', 'ArtistController@show');
-    Route::patch('/{artist}', 'ArtistController@update');
-    Route::delete('/{artist}', 'ArtistController@destroy');
+    Route::post('', 'ArtistController@store')->middleware('permission:store artist');
+    Route::get('', 'ArtistController@index')->middleware('permission:get artist');
+    Route::get('/{artist}', 'ArtistController@show')->middleware('permission:get artist');
+    Route::patch('/{artist}', 'ArtistController@update')->middleware('permission:update artist');
+    Route::delete('/{artist}', 'ArtistController@destroy')->middleware('permission:destroy artist');
 });
 
 Route::group(["namespace" => 'Api', 'prefix' => 'customer', 'middleware' => 'auth:api'], function () {
-    Route::post('', 'CustomerController@store');
-    Route::get('', 'CustomerController@index');
-    Route::get('/{customer}', 'CustomerController@show');
-    Route::patch('/{customer}', 'CustomerController@update');
-    Route::delete('/{customer}', 'CustomerController@destroy');
+    Route::post('', 'CustomerController@store')->middleware('permission:store customer');
+    Route::get('', 'CustomerController@index')->middleware('permission:get customer');
+    Route::get('/{customer}', 'CustomerController@show')->middleware('permission:get customer');
+    Route::patch('/{customer}', 'CustomerController@update')->middleware('permission:update customer');
+    Route::delete('/{customer}', 'CustomerController@destroy')->middleware('permission:destroy customer');
 });
 
 Route::group(["namespace" => 'Api', 'prefix' => 'order', 'middleware' => 'auth:api'], function () {
@@ -56,14 +59,14 @@ Route::group(["namespace" => 'Api', 'prefix' => 'order', 'middleware' => 'auth:a
 });
 
 Route::group(["namespace" => 'Api', 'prefix' => 'artwork', 'middleware' => 'auth:api'], function () {
-    Route::post('', 'ArtworkController@store');
-    Route::get('', 'ArtworkController@index');
-    Route::get('/{artwork}', 'ArtworkController@show');
-    Route::patch('/{artwork}', 'ArtworkController@update');
-    route::delete('/{artwork}', 'ArtworkController@destroy');
+    Route::post('', 'ArtworkController@store')->middleware('permission:store artwork');
+    Route::get('', 'ArtworkController@index')->middleware('permission:get artwork');
+    Route::get('/{artwork}', 'ArtworkController@show')->middleware('permission:get artwork');
+    Route::patch('/{artwork}', 'ArtworkController@update')->middleware('permission:update artwork');
+    route::delete('/{artwork}', 'ArtworkController@destroy')->middleware('permission:destroy artwork');
 
-    Route::post('/{artwork}/image', 'ArtworkController@storeImage');
-    Route::delete('/{artwork}/image/{media}', 'ArtworkController@destroyImage');
+    Route::post('/{artwork}/image', 'ArtworkController@storeImage')->middleware('permission:store artwork image');
+    Route::delete('/{artwork}/image/{media}', 'ArtworkController@destroyImage')->middleware('permission:destroy artwork image');
 });
 
 Route::group(["namespace" => 'Api', 'prefix' => 'order', 'middleware' => 'auth:api'], function () {
